AMP final

+++ AMP_C_arm_2/utils/clock.c
@@ -2,7 +2,7 @@
 #include "clock.h"
 
 
-unsigned int getTimeshtamp() {
+unsigned long long getTimeshtamp() {
 
     unsigned cc;
